refactor(stores): eliminate delegation duplication via shared base classes
- Add _DelegatingMemoryStore and _DelegatingGraphStore to _store_utils with
  typed _MemoryBackend and _GraphBackend Protocols to satisfy mypy
- DuckDB/PostgreSQL/MySQL/MongoDB/Redis stores inherit bases; __init__-only
- Add neighbors() alias on _RelationalBackend; widen put_memories/upsert_edges
  signatures in _MongoBackend and _RedisBackend from tuple to Sequence
- Replace _InMemoryMemory/GraphStore stubs in evals/runner.py and dreaming
  tests with production InMemoryMemoryStore/InMemoryGraphStore imports
- Extract shared _memory/_edge factories and assert_list_by_filtering helper
  to tests/integration/stores/_helpers.py; use from SQL and cache test files
- Add pythonpath=["."] to pytest config to enable cross-test-package imports
